Create asic financial analysis on a sample data set:
In this example, we assume that the financial data is in a CSV file called financial_data.csv and that it contains columns for Earnings, Shares, Price, Equity, and Sector. The script calculates the EPS, PE Ratio, Book Value, and PB Ratio for each company, and then calculates the average ratios for the sector. The script then joins the sector data back to the financial data and calculates the relative ratios to the sector averages. The results are output to a CSV file called financial_analysis.csv.

Again, please note that this is just an example script, and you will need to modify it based on the specific data and analysis that you want to perform. Additionally, you will need to ensure that the financial data you are using is accurate and up-to-date.